tweak: don't kill players

+++ Code/client/Services/Generic/ActorValueService.cpp
@@ -292,7 +292,12 @@ void ActorValueService::OnHealthChangeBroadcast(const NotifyHealthChangeBroadcas
 
     const float health = pActor->GetActorValue(ActorValueInfo::kHealth);
     if (health <= 0.f)
-        pActor->Kill();
+    {
+        ActorExtension* pExtension = pActor->GetExtension();
+        // Players should never be killed
+        if (!pExtension->IsPlayer())
+            pActor->Kill();
+    }
 }
 
 void ActorValueService::OnActorValueChanges(const NotifyActorValueChanges& acMessage) const noexcept
@@ -378,6 +383,11 @@ void ActorValueService::OnDeathStateChange(const NotifyDeathStateChange& acMessa
     Actor* pActor = Cast<Actor>(TESForm::GetById(formIdComponent.Id));
 
     if (!pActor)
+        return;
+
+    ActorExtension* pExtension = pActor->GetExtension();
+    // Players should never be killed
+    if (pExtension->IsPlayer())
         return;
 
     if (pActor->IsDead() != acMessage.IsDead)
